+++ src/cli.ts
@@ -47,6 +47,7 @@ import {
 
 import { getForgeVersion } from "./util/version.js";
 import { shellCompletionScript } from "./util/completion-script.js";
+import { providerTimeoutMs } from "./util/abort.js";
 import {
   pruneToolOutputsSync,
   toolOutputStats,